Slash commands fix (#3)
* Use new slash commands API

Switches to the new slash commands API, as the old one is no longer functional.

* Use RN headers for the application index request

* Use Patcher.addPatch instead of patcher.patch

* Remove debug prints

* Update spoofed RN version

* Fix spoofed RN version being latest Alpha instead of Beta

* Use an enum indicating the source of the request for passCommandData

* Reimplement handleApplicationCommandsUpdate logic

* Fix commands not popping up immediately when browsing

The issue was caused by the requestApplicationCommands patch checking for the
application ID instead of guild ID and thus always being bypassed.

* Use GsonUtils to get the Gson object

* Remove cachedGuilds in favor of guildApplicationIndexes

* Use custom classes for deserialization

Fixes the crash with guilds that have permissions set up for application
commands. Currently completely ignores permissions. Also fixes some minor issues
due to the substantial refactoring caused by this.

* Remove old import

* Remove debug print

* Fix builtin commands not working

* Fix application ordering when browsing commands

* Add newlines after hook variable setup

* Add a way to handle command permissions

Doesn't contain an actual implementation of command permissions, just the base
work to make them work. This is a lot of code anyways.

* Implement command permissions

* Remove a debug print

* Add a newline

* Sort imports

* Fix a null pointer exception with default permissions

If the API provides default (empty) permissions, represent it appropriately in
the model, instead of using null.

* Improve the permission checking code

* Allow guild owners to use all commands

* Always allow administrators to use commands

* Remove commented out code

* Fix null dereference if not all permissions are sent by API

Permission checking assumes that the Permissions object doesn't have null
fields, but the API doesn't always send all fields for it. This defaults the
null fields when constructing the object.

* Rewrite if chain as switch

* Fix valid unchecked warnings and ignore invalid ones

* Fix autocomplete of builtin commands

* Add a newline

* Split classes to separate files

* Allow requesting either the guild or DM endpoints

* Fix commands not working in DMs

* Add a return statement for InsteadHook

* Add newlines before returns

* Fix guild commands not working

* Null out defaultPermissions as it's not used

* Prefer bot user name and icon over application name and icon

* Consider guild_member_permissions when checking permissions

* Invalidate cache on GUILD_APPLICATION_COMMAND_INDEX_UPDATE

* Add missing source file

* Rename GUILD request source to INITIAL

It's not only used for guilds anymore, so rename it to avoid confusion.

* Invalidate cache on invalid command version

* Sort imports

* Display a toast after using an outdated command

* Trigger cache invalidation only on INTERACTION_APPLICATION_COMMAND_INVALID_VERSION

* Update TODO comments

* Use Optional return for ApplicationIndexSource.getIndex

* Move caches to a separate class

* Remove unnecessary imports

* Fix crash on guild change

When switching from guild A to guild B, Discord checks for permissions of
commands from guild A, even though they are already replaced with ones from
guild B. This was causing a lot of crashes in the hasPermission hook.

This fixes this by simply checking if the application we are checking permission
for actually exists in the guild. This is probably better done in another way,
but hey, it works for now.

* Use sensible names for magic values

* Shorten Comparator to a lambda

* Make command permissions actually act as overrides

Discord doesn't call command permissions *overrides* for nothing - they override
the application permissions. Until now, I was checking if both application
permissions and command permissions allow for the command to be used, this
instead checks if either allow for it.

* Make permission checking logic more readable

* Fix deny permission overrides not applying

* Fix outdated command handling

* Display server-provided error message instead of a hardcoded one

This seems to be what the official client does, and it fixes localization.

* Remove accidentally added file

* Make ApplicationIndex.applications a Map instead of List

* Make ApplicationIndex.applicationCommands a Map instead of List

* Implement user applications

* Cache methods

* Calculate command counts dynamically

Fixes an edge case with user apps.

* Fix builtin commands not loading

Regression introduced in efade6cd2763f96e78184b327e6b30eadefef02a.

* Fix exception when trying to use commands in non-bot DM

Commands in non-bot DMs still don't appear, but they don't outright
crash anymore.

* Load commands lazily

This also fixes commands in non-bot DMs.

* Filter out non-slash commands

* Fix permission check in threads

Threads inherit the permissions of their parent channels, but we were
checking for the permissions of the thread itself (which do not exist).
This caused disallowed commands to not work in threads of channels that
had overrides for them.

* Replace some magic values with constants

* Remove commented out code

* Fix commands not working in some DMs

isBot is sometimes null and causes an exception. Now assume it's false
in those cases. Seems to be the proper way, though not sure why it
mostly worked without this - from my own tests Discord always skips the
bot field unless it's true.

* Add comment to hook

* Fix completion not working in DMs

For some reason there was different handling for generating index
sources between browsing and completing commands.

* Correct comment

* Reorder modifiers

* Cache fields for reflection

---------

Co-authored-by: Grzesiek11 <grzesiek11@stary.pc.pl>